third_party/WebKit/LayoutTests/css3/selectors3/xml/css3-modsel-15c-expected.txt

CONSOLE ERROR: line 9: Uncaught TypeError: document.getElementsByTagNameNS(...)[0].setIdAttribute is not a function
